# Add v8-compile-cache to improve startup performance