--- a/package.json
+++ b/package.json
@@ -45,6 +45,7 @@
     "is-valid-glob": "^1.0.0",
     "micromatch": "^4.0.2",
     "resolve": "^1.17.0",
+    "v8-compile-cache": "^2.1.1",
     "yargs": "^15.4.1"
   },
   "devDependencies": {