Running the ember-template-lint binary directly with no args is 9% faster with [v8-compile-cache](https://www.npmjs.com/package/v8-compile-cache):

```
~/Development/ember-template-lint hyperfine 'DISABLE_V8_COMPILE_CACHE=1 ./bin/ember-template-lint.js' './bin/ember-template-lint.js' --ignore-failure
Benchmark #1: DISABLE_V8_COMPILE_CACHE=1 ./bin/ember-template-lint.js
  Time (mean ± σ):     159.6 ms ±   3.1 ms    [User: 148.4 ms, System: 28.3 ms]
  Range (min … max):   153.7 ms … 165.5 ms    17 runs

  Warning: Ignoring non-zero exit code.

Benchmark #2: ./bin/ember-template-lint.js
  Time (mean ± σ):     146.2 ms ±   2.4 ms    [User: 135.1 ms, System: 28.7 ms]
  Range (min … max):   141.6 ms … 150.1 ms    19 runs

  Warning: Ignoring non-zero exit code.

Summary
  './bin/ember-template-lint.js' ran
    1.09 ± 0.03 times faster than 'DISABLE_V8_COMPILE_CACHE=1 ./bin/ember-template-lint.js'
```

ESLint uses this too: https://github.com/eslint/eslint/pull/11921

Do we think this small startup time savings is worthwhile?